Added support for matching multiple users to the same match

+++ server.py
@@ -23,6 +23,7 @@ class GameRequest(GameDataBaseModel):
 	latitude = DecimalField(9,6)
 	name = CharField(20)
 	game = CharField(20)
+	available = IntegerField()
 
 class User(GameDataBaseModel):
 	userid = IntegerField()
@@ -31,27 +32,28 @@ class User(GameDataBaseModel):
 # definitions for different types of request from the user
 
 IDENTIFY 	= 0
-GAMEREQUEST = 1
+GAMEREQUEST	= 1
 SENDDATA	= 2
 NEWUSER		= 3
 RESET		= 4
 
 #definitions for communication codes to send back to user
 
-ERRORCODE			= 0
+ERRORCODE		= 0
 
 IDENTIFYCODEOK		= 1
 IDENTIFYCODEFAIL	= 2
 
-NEWUSEROK			= 3
-NEWUSERFAIL			= 4
+NEWUSEROK		= 3
+NEWUSERFAIL		= 4
 
-REQUESTWAIT			= 5
+REQUESTWAIT		= 5
 REQUESTFOUND		= 6
 
-DATACODE			= 7
-LEFTCODE			= 8
+DATACODE		= 7
+LEFTCODE		= 8
 
+MATCHSTART		= 9
 
 
 # objects to keep track of users, and allow communication between them
@@ -62,6 +64,7 @@ users = {}
 
 GameRequest.delete()
 
+# start protocal for handling incoming connections
 
 class Game(protocol.Protocol):
 	def connectionMade(self):
@@ -108,9 +111,9 @@ class Game(protocol.Protocol):
 		# FIX generally nothing is communicated back to client
 		
 		# IDENTIFY USER		0:username:
-		# GAME REQUEST		1:game:longitude:latitude:tolerance:
+		# GAME REQUEST		1:game:longitude:latitude:tolerance:openspots:
 		# SEND DATA			2:data
-		# NEW USER			3:username
+		# NEW USER			3:username:
 		# LEAVE				4:
 		
 		command = int(data[0])
@@ -121,9 +124,9 @@ class Game(protocol.Protocol):
 			self.identifyUser(username)
 			
 		if (command == GAMEREQUEST):
-			cmd,game,longitude,latitude,tolerance,endline = data.split(":")
+			cmd,game,longitude,latitude,tolerance,availabile,endline = data.split(":")
 			print "Game request for: " + game + " At: " + longitude + " -- " + latitude
-			self.gameRequest(game,float(longitude),float(latitude),float(tolerance))
+			self.gameRequest(game,float(longitude),float(latitude),float(tolerance),int(available))
 			
 		if (command == SENDDATA):
 			self.sendData(data[2:])
@@ -172,7 +175,7 @@ class Game(protocol.Protocol):
 			# save the new user, still have to call the identify user code as usual
 			self.userCommunicateCode(NEWUSEROK)
 			
-	def gameRequest(self,game,longitude,latitude,tolerance):
+	def gameRequest(self,game,longitude,latitude,tolerance,available):
 		# now we can either find a request that's already there, or make one to wait
 		# FIRST look for requests that match the one that we already have
 		tolerance = tolerance/70 # This is a really rough hack to get approximate longitude/latitude difference in miles
@@ -184,21 +187,34 @@ class Game(protocol.Protocol):
 			(GameRequest.longitude < (longitude + tolerance)) & 
 			(GameRequest.longitude > (latitude - tolerance)) ) \
 			.order_by(GameRequest.id) # this is so we fairly choose the oldest request
+			# it shouldn't matter if a later player has a larger tolerance than us, because they'll travel
 		
 		if matchingRequests.count() != 0:
 			# We found one or more matches, so pick the first one of them, the oldest
 			selectedRequest = matchingRequests[0]
 			print "Found match for this game, with user: ",selectedRequest.name
-			selectedRequest.delete_instance() #we can remove the request from the database
 			
 			#now we need to tell the Game processes about each other
-			self.users[selectedRequest.name] = users[selectedRequest.name]
-			users[selectedRequest.name].users[self.user.name] = self
+			self.users[selectedRequest.name] = users[selectedRequest.name] 	#copy their object into our list of users in match
+
+			for userkey in users[selectedRequest.name].users:		#copy our name into all existing members in match
+				users[selectedRequest.name].users[userkey].users[self.user.name] = self
+				users[selectedRequest.name].userCommunicateCode(REQUESTFOUND)
 			
-			users[selectedRequest.name].userCommunicateCode(REQUESTFOUND)
+			users[selectedRequest.name].users[self.user.name] = self	#copy our name into their list of users in match
 			self.userCommunicateCode(REQUESTFOUND)
 			
-			users.pop(selectedRequest.name, None) # Remove them from the list of waiting users
+			#handle removing requests:
+			selectedRequest.available = selectedRequest.available - 1
+			selectedRequest.save()
+
+			if selectedRequest.available == 0:
+				selectedRequest.delete_instance() #we can remove the request from the database
+				for userkey in users[selectedRequest.name].users:
+					users[selectedRequest.name].users[userkey].userCommunicateCode(MATCHSTART)
+				users[selectedRequest.name].userCommunicateCode(MATCHSTART)
+
+				users.pop(selectedRequest.name, None) # Remove them from the list of waiting users
 			
 		else:
 			# We didn't find any matches, so we need to create a new GameRequest for this request and add it.
@@ -208,10 +224,11 @@ class Game(protocol.Protocol):
 			newgame.latitude = latitude
 			newgame.name = self.user.name
 			newgame.game = game
+			newgame.available = available
 			newgame.save()
 			self.request = newgame
 			
-			print "No match found, created new match",newgame.longitude,newgame.latitude,self.user.name,game,newgame.id
+			print "No match found, created new match",newgame.longitude,newgame.latitude,self.user.name,game,newgame.id,available
 			
 			# add this user's Game object to the list of waiting users, so we can communicate back and forth
 			# when someone else finds us for a match
